+++ crates/owl-ms-language-server/Cargo.toml
@@ -32,6 +32,7 @@ walkdir = "2.5.0"
 
 [dev-dependencies]
 criterion = { version = "0.4", features = ["html_reports"] }
+pretty_assertions = "1"
 
 [build-dependencies]
 cc = "1.0"
